refactor: layout 타입 적용
